fix: Fixes

+++ re-deployer/re-deployer.js
@@ -2,15 +2,16 @@ const { execSync } = require("child_process");
 const fs = require("fs");
 const path = require("path");
 
-const TIME_20_SECONDS = 20 * 1000;
-const TIME_20_MINUTES = 20 * 60 * 1000;
 const TIME_1_SECOND = 1000;
 const TIME_5_SECONDS = 5 * 1000;
 const TIME_10_SECONDS = 10 * 1000;
+const TIME_20_SECONDS = 20 * 1000;
+const TIME_5_MINUTES = 5 * 60 * 1000;
+const TIME_30_MINUTES = 30 * 60 * 1000;
 
 const DEV_MODE = !!process.env.DEV_MODE || false;
-const DEPLOY_BLOCK_DURATION = DEV_MODE ? TIME_20_SECONDS : TIME_20_MINUTES;
-const ZIP_TRANSFER_BLOCK_TIME = TIME_10_SECONDS;
+const DEPLOY_BLOCK_DURATION = DEV_MODE ? TIME_20_SECONDS : TIME_5_MINUTES;
+const ZIP_TRANSFER_BLOCK_TIME = DEV_MODE ? TIME_1_SECOND : TIME_10_SECONDS;
 const DEPARTURE_LOUNGE_LOCATION = DEV_MODE
   ? "/Users/george.gillams/Documents/departure-lounge"
   : "/home/ubuntu/departure-lounge";
@@ -24,10 +25,13 @@ const SSL_KEY_PATH = DEV_MODE
   ? "/Users/george.gillams/Documents/letsencrypt/live"
   : "/etc/letsencrypt/live";
 const CHECK_FREQUENCY = DEV_MODE ? TIME_1_SECOND : TIME_5_SECONDS;
+const MAX_UNZIP_ATTEMPT_TIME = DEV_MODE ? TIME_20_SECONDS : TIME_30_MINUTES;
 
-const ZIP_FORMAT = /[a-zA-Z0-9\.]+\-\-\-[a-zA-Z0-9]+\.zip/gi;
+const ZIP_FORMAT = /[a-zA-Z0-9\.]+\-\-\-[a-zA-Z0-9\-_]+\.zip/gi;
 
 let deploysBlockedUntil = 0;
+
+const unzipAttemptTime = {};
 
 const calculatePaths = (fileName) => {
   let fileNameWOExt = fileName.endsWith(".zip")
@@ -55,16 +59,12 @@ function getNginxConfigText(site, ssl) {
     site.server_name
   }`;
 
-  const sslConfig = [
-    `ssl_dhparam "/etc/pki/nginx/dhparams.pem";`,
-    `add_header Strict-Transport-Security "max-age=31536000; includeSubDomains" always;`,
-    `ssl_certificate /etc/letsencrypt/live/${site.server_name}/fullchain.pem;`,
-    `ssl_certificate_key /etc/letsencrypt/live/${site.server_name}/privkey.pem;`,
-    `include /etc/letsencrypt/options-ssl-nginx.conf;`,
-  ];
-  const sslBlock = sslConfig
-    .map((line) => `${ssl ? "  " : "  # "}${line}`)
-    .join(`\n`);
+  const sslKeyConfig = `  ssl_dhparam "/etc/pki/nginx/dhparams.pem";
+  add_header Strict-Transport-Security "max-age=31536000; includeSubDomains" always;
+  ssl_certificate /etc/letsencrypt/live/${site.server_name}/fullchain.pem;
+  ssl_certificate_key /etc/letsencrypt/live/${site.server_name}/privkey.pem;
+  include /etc/letsencrypt/options-ssl-nginx.conf;
+  `;
 
   let config = `server {
   server_name ${site.server_name} www.${site.server_name};
@@ -73,10 +73,11 @@ function getNginxConfigText(site, ssl) {
   listen 80${site.is_default_site ? " default_server" : ""};
   listen [::]:80${site.is_default_site ? " default_server" : ""};
 
+  ${
+    ssl
+      ? `
   # Redirect non-https traffic to https
-  if ($scheme != "https") {
-    return 301 https://$host$request_uri;
-  }
+  return 301 https://$host$request_uri;
 }
 
 server {
@@ -86,7 +87,7 @@ server {
   listen 443 ssl http2;
   listen [::]:443 ssl http2;
 
-${sslBlock}
+${sslKeyConfig}
 
   return 301 https://${domainRedirectTo}$request_uri;
 }
@@ -94,6 +95,14 @@ ${sslBlock}
 server {
   server_name ${domainRedirectTo};
   root         /usr/share/nginx/html;
+
+  listen 443 ssl http2${site.is_default_site ? " default_server" : ""};
+  listen [::]:443 ssl http2${site.is_default_site ? " default_server" : ""};
+
+${sslKeyConfig}
+  `
+      : ``
+  }
 
   error_page   404  /404.html;
   location = /404.html {
@@ -105,13 +114,8 @@ server {
       root   /var/www/html/errors;
   }
 
-  listen 443 ssl http2${site.is_default_site ? " default_server" : ""};
-  listen [::]:443 ssl http2${site.is_default_site ? " default_server" : ""};
-
-${sslBlock}
-
   location / {
-    proxy_pass http://127.0.0.1:${site.server_port};
+    proxy_pass http://127.0.0.1:${site.host_port};
     proxy_set_header        X-Real-IP       $remote_addr;
     proxy_set_header        Host            $host;
     proxy_redirect          off;
@@ -139,7 +143,7 @@ const generateNginxConfig = (site, withSsl) => {
   fs.writeFileSync(outputPath, text);
   execSync(`ln -s ${outputPath} ${NGINX_SITES_ENABLED_LOCATION}/ || true`);
   if (!DEV_MODE) {
-    execSync("service nginx restart");
+    execSync("sudo service nginx restart");
   }
 };
 
@@ -158,21 +162,38 @@ const configureSSL = (serverName) => {
     return;
   }
   execSync(
-    `/usr/bin/certbot-auto certonly --nginx -d ${serverName} -d www.${serverName} --debug`
+    `sudo certbot certonly --nginx -d ${serverName} -d www.${serverName} --debug`
   );
 };
 
-const abortDeploy = (filePath, extractionPath) => {
+const cleanup = (filePath, extractionPath) => {
   for (let path of [filePath, extractionPath]) {
-    if (pathsExistsSimple(path)) {
+    if (path && pathsExistsSimple(path)) {
       fs.rmSync(path, { force: true, recursive: true });
     }
   }
 };
 
-const extractZip = (filePath, extractionPath) => {
-  execSync(`unzip -o ${filePath} -d ${DEPARTURE_LOUNGE_LOCATION}`);
+const extractZip = (filePath, extractionPath, fileNameWOExt) => {
+  if (pathsExistsSimple(extractionPath)) {
+    throw new Error(
+      `Extraction location ${fileNameWOExt} is already in use. To redeploy, send a file with a unique hash.`
+    );
+  }
+  execSync(`mkdir -p ${extractionPath}`);
+  execSync(`unzip -o ${filePath} -d ${extractionPath}`);
   execSync(`rm -rf ${filePath}`);
+};
+
+const checkKeysExist = (meta, expectedKeys, fileNameWOExt) => {
+  const metaKeys = Object.keys(meta);
+  for (let expectedKey of expectedKeys) {
+    if (!metaKeys.includes(expectedKey)) {
+      throw new Error(
+        `Cannot deploy app ${fileNameWOExt} as meta data ${expectedKey} is not defined`
+      );
+    }
+  }
 };
 
 const readMetaFile = (metaPath, fileNameWOExt) => {
@@ -183,14 +204,12 @@ const readMetaFile = (metaPath, fileNameWOExt) => {
   }
 
   const meta = JSON.parse(fs.readFileSync(metaPath, "utf8"));
-  const metaKeys = Object.keys(meta);
-  for (let expectedKey of ["server_port", "server_name", "redirect_to_www"]) {
-    if (!metaKeys.includes(expectedKey)) {
-      throw new Error(
-        `Cannot deploy app ${fileNameWOExt} as meta data ${expectedKey} is not defined`
-      );
-    }
+  checkKeysExist(meta, ["is_webapp"], fileNameWOExt);
+  const expectedKeys = ["is_webapp", "host_port", "docker_port", "server_name"];
+  if (meta.is_webapp) {
+    expectedKeys.push("redirect_to_www");
   }
+  checkKeysExist(meta, expectedKeys, fileNameWOExt);
 
   return meta;
 };
@@ -199,16 +218,40 @@ const sslAlreadyConfigured = (serverName) => {
   return pathsExistsSimple(path.join(SSL_KEY_PATH, serverName, "privkey.pem"));
 };
 
-const generatePm2Config = (outputPath, site, dockerImage) => {
+const generatePm2Config = (outputPath, dockerImage, dockerContainerId) => {
   const text = `module.exports = {
   apps: [
     {
       name: '${dockerImage}',
-      script: 'docker run -t -p ${site.server_port}:3000 ${dockerImage}',
+      script: 'docker start ${dockerContainerId}',
+      restart_delay: 10000,
     },
   ],
 };`;
   fs.writeFileSync(outputPath, text);
+};
+
+const getDockerImages = () => {
+  const dockerImagesJson = execSync(
+    "curl --unix-socket /var/run/docker.sock http://localhost/v1.24/images/json?all=1"
+  );
+
+  const dockerImageData = JSON.parse(dockerImagesJson);
+  return dockerImageData;
+};
+const getDockerContainers = () => {
+  const dockerContainersJson = execSync(
+    "curl --unix-socket /var/run/docker.sock http://localhost/v1.24/containers/json?all=1"
+  );
+
+  const dockerContainerData = JSON.parse(dockerContainersJson);
+  return dockerContainerData;
+};
+const getPm2Processes = () => {
+  const pm2JsonListOutput = execSync(`pm2 jlist`).toString().split("\n");
+  const pm2JsonList = pm2JsonListOutput[pm2JsonListOutput.length - 1];
+  const pm2Data = JSON.parse(pm2JsonList);
+  return pm2Data;
 };
 
 const getOldAppItems = (list, appName, newHash) => {
@@ -218,27 +261,24 @@ const getOldAppItems = (list, appName, newHash) => {
   );
 };
 
+const getMatchingDockerImage = (appName, hash) => {
+  return getDockerImages().find(
+    (entry) => entry.RepoTags[0].split(":latest")[0] === `${appName}---${hash}`
+  );
+};
 const getOldDockerImages = (appName, newHash) => {
-  const dockerImageEntries = execSync("docker image list --all")
-    .toString()
-    .split("\n")
-    .filter(
-      (entry) =>
-        entry.includes(`${appName}---`) &&
-        !entry.includes(`${appName}---${newHash}`)
-    );
-  return dockerImageEntries.map((entry) => entry.split(" ")[0]);
+  const dockerImageNames = getDockerImages().map(
+    (entry) => entry.RepoTags[0].split(":latest")[0]
+  );
+  return getOldAppItems(dockerImageNames, appName, newHash);
 };
 const getOldDockerContainers = (appName, newHash) => {
-  const dockerContainerEntries = execSync("docker container list --all")
-    .toString()
-    .split("\n")
-    .filter(
-      (entry) =>
-        entry.includes(`${appName}---`) &&
-        !entry.includes(`${appName}---${newHash}`)
-    );
-  return dockerContainerEntries.map((entry) => entry.split(" ")[0]);
+  const dockerContainerEntries = getDockerContainers().filter(
+    (entry) =>
+      entry.Image.startsWith(`${appName}---`) &&
+      !entry.Image.endsWith(`---${newHash}`)
+  );
+  return dockerContainerEntries.map((entry) => entry.Id);
 };
 const getOldDeployDirectories = (appName, newHash) => {
   const fileList = execSync(`ls ${DEPARTURE_LOUNGE_LOCATION}`)
@@ -247,9 +287,7 @@ const getOldDeployDirectories = (appName, newHash) => {
   return getOldAppItems(fileList, appName, newHash);
 };
 const getOldPm2Processes = (appName, newHash) => {
-  const pm2List = JSON.parse(execSync(`pm2 jlist`).toString()).map(
-    (entry) => entry.name
-  );
+  const pm2List = getPm2Processes().map((entry) => entry.name);
   return getOldAppItems(pm2List, appName, newHash);
 };
 
@@ -277,20 +315,56 @@ const deploy = (fileName) => {
       hash,
     });
 
-    extractZip(filePath, extractionPath);
+    try {
+      extractZip(filePath, extractionPath, fileNameWOExt);
+    } catch (error) {
+      if (
+        error
+          .toString()
+          .includes("End-of-central-directory signature not found.")
+      ) {
+        if (unzipAttemptTime[fileNameWOExt] > MAX_UNZIP_ATTEMPT_TIME) {
+          throw new Error(
+            `Zip file has repeatedly failed. Removing incomplete file.`
+          );
+        }
+        console.log(`Zip file is not yet fully uploaded.`);
+        cleanup(extractionPath);
+        deploysBlockedUntil = Date.now() + ZIP_TRANSFER_BLOCK_TIME;
+        unzipAttemptTime[fileNameWOExt] =
+          (unzipAttemptTime[fileNameWOExt] || 0) + ZIP_TRANSFER_BLOCK_TIME;
+        return;
+      } else {
+        throw error;
+      }
+    }
 
-    const meta = readMetaFile(metaPath);
+    const meta = readMetaFile(metaPath, fileNameWOExt);
     console.table(meta);
 
-    let sslConfigured = sslAlreadyConfigured(meta.server_name);
-    if (!sslConfigured) {
+    if (meta.is_webapp) {
+      let sslConfigured = sslAlreadyConfigured(meta.server_name);
+      if (!sslConfigured) {
+        generateNginxConfig(meta, sslConfigured);
+        try {
+          configureSSL(meta.server_name);
+          sslConfigured = true;
+        } catch (error) {
+          console.log(
+            `Error configuring SSL. This should be done manually and the project redeployed.`
+          );
+        }
+      }
       generateNginxConfig(meta, sslConfigured);
-      configureSSL(meta.server_name);
-      sslConfigured = true;
     }
-    generateNginxConfig(meta, sslConfigured);
 
-    generatePm2Config(pm2ConfigPath, meta, fileNameWOExt);
+    const duplicateImage = getMatchingDockerImage(appName, hash);
+    if (duplicateImage) {
+      throw new Error(
+        `An image for app ${appName} with hash ${hash} is already loaded in Docker. To redeploy, send an image with a unique hash.`
+      );
+    }
+
     // Load to docker
     execSync(`docker load < ${path.join(extractionPath, "docker-image.tar")}`);
 
@@ -319,6 +393,15 @@ const deploy = (fileName) => {
       execSync(`docker image rm --force ${dockerImage}`);
     });
 
+    // add new docker image
+    const dockerContainerId = execSync(
+      `docker create -t -p ${meta.host_port}:${meta.docker_port} ${fileNameWOExt}`
+    )
+      .toString()
+      .split("\n")[0];
+
+    generatePm2Config(pm2ConfigPath, fileNameWOExt, dockerContainerId);
+
     // Run with PM2
     execSync(`pm2 start ${pm2ConfigPath}`);
     // PM2 save
@@ -329,16 +412,9 @@ const deploy = (fileName) => {
       execSync(`rm -rf ${path.join(DEPARTURE_LOUNGE_LOCATION, directory)}`);
     });
   } catch (error) {
-    if (
-      error.toString().includes("End-of-central-directory signature not found.")
-    ) {
-      console.log(`Zip file is not yet fully uploaded.`);
-      deploysBlockedUntil = Date.now() + ZIP_TRANSFER_BLOCK_TIME;
-      return;
-    }
-
     console.error(`Error deploying: ${error}`);
-    abortDeploy(filePath, extractionPath);
+    delete unzipAttemptTime[fileNameWOExt];
+    cleanup(filePath, extractionPath);
   }
 };
